Basic Editor

+++ Lumos/external/imgui/plugins/ImCurveEdit.cpp
@@ -194,8 +194,8 @@ namespace ImCurveEdit
                     
                     if (distance(io.MousePos.x, io.MousePos.y, pos1.x, pos1.y, pos2.x, pos2.y) < 8.f && localOverCurve == -1)
                     {
-                        localOverCurve = c;
-                        overCurve = c;
+                        localOverCurve = static_cast<int>(c);
+                        overCurve = static_cast<int>(c);
                         overCurveOrPoint = true;
                     }
                     
@@ -261,7 +261,7 @@ namespace ImCurveEdit
             {
                 for (size_t p = 0; p < ptCount; p++)
                 {
-                    delegate.EditPoint(movingCurve, p, rangeToPoint(pointToRange(pts[p]) + io.MouseDelta * sizeOfPixel));
+                    delegate.EditPoint(movingCurve, static_cast<int>(p), rangeToPoint(pointToRange(pts[p]) + io.MouseDelta * sizeOfPixel));
                 }
             }
             if (!io.MouseDown[0])
